requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.7.0
alembic>=1.13.0
pyyaml>=6.0.0
